requirements.txt
mysqlclient
stripe
Unidecode
Pillow
django-resized
